requirements.txt
canal-python==0.4
mysqlclient==2.0.3